AddMatchForm moved with its tests to its own folder
